fix(类型): 修复 SQL 查询参数占位符和类型转换
- 在 finance.rs 的 payables 和 receivables 查询中添加 ::BIGINT 类型转换，确保 SUM 聚合函数返回 i64 类型
- 在 reports.rs 中修复 SQL 参数占位符编号错误（$11、$21 等改为 $1、$2 等），确保参数正确绑定
- 修复 inventory_report_summary 查询中的参数占位符不一致问题
- 修复 inventory_aging_analysis 查询中的参数占位符重复和错误编号
- 修复 inventory_slow_moving 查询中的参数占位符编号错误
- 这些修复确保 SQL 查询能够正确接收和处理绑定参数，避免运行时错误

+++ src-tauri/src/commands/finance.rs
@@ -202,12 +202,12 @@ pub async fn get_payables(
     let summary = sqlx::query_as::<_, (i64, i64, i64, i64)>(
         r#"
         SELECT
-            COALESCE(SUM(CASE WHEN adjustment_type = 'normal' THEN payable_amount ELSE 0 END), 0),
-            COALESCE(SUM(CASE WHEN status = 'paid' AND adjustment_type = 'normal' THEN payable_amount ELSE 0 END), 0),
-            COALESCE(SUM(CASE WHEN status = 'partial' AND adjustment_type = 'normal' THEN payable_amount ELSE 0 END), 0),
+            COALESCE(SUM(CASE WHEN adjustment_type = 'normal' THEN payable_amount ELSE 0 END), 0)::BIGINT,
+            COALESCE(SUM(CASE WHEN status = 'paid' AND adjustment_type = 'normal' THEN payable_amount ELSE 0 END), 0)::BIGINT,
+            COALESCE(SUM(CASE WHEN status = 'partial' AND adjustment_type = 'normal' THEN payable_amount ELSE 0 END), 0)::BIGINT,
             COALESCE(SUM(CASE WHEN status != 'paid' AND adjustment_type = 'normal'
                 AND due_date IS NOT NULL AND due_date < CURRENT_DATE::TEXT
-                THEN (payable_amount - paid_amount) ELSE 0 END), 0)
+                THEN (payable_amount - paid_amount) ELSE 0 END), 0)::BIGINT
         FROM payables
         "#,
     )
@@ -510,12 +510,12 @@ pub async fn get_receivables(
     let summary = sqlx::query_as::<_, (i64, i64, i64, i64)>(
         r#"
         SELECT
-            COALESCE(SUM(CASE WHEN adjustment_type = 'normal' THEN receivable_amount ELSE 0 END), 0),
-            COALESCE(SUM(CASE WHEN status = 'paid' AND adjustment_type = 'normal' THEN receivable_amount ELSE 0 END), 0),
-            COALESCE(SUM(CASE WHEN status = 'partial' AND adjustment_type = 'normal' THEN receivable_amount ELSE 0 END), 0),
+            COALESCE(SUM(CASE WHEN adjustment_type = 'normal' THEN receivable_amount ELSE 0 END), 0)::BIGINT,
+            COALESCE(SUM(CASE WHEN status = 'paid' AND adjustment_type = 'normal' THEN receivable_amount ELSE 0 END), 0)::BIGINT,
+            COALESCE(SUM(CASE WHEN status = 'partial' AND adjustment_type = 'normal' THEN receivable_amount ELSE 0 END), 0)::BIGINT,
             COALESCE(SUM(CASE WHEN status != 'paid' AND adjustment_type = 'normal'
                 AND due_date IS NOT NULL AND due_date < CURRENT_DATE::TEXT
-                THEN (receivable_amount - received_amount) ELSE 0 END), 0)
+                THEN (receivable_amount - received_amount) ELSE 0 END), 0)::BIGINT
         FROM receivables
         "#,
     )
